CodeFix

+++ ARTChat/Src/SocketServer/ConnectionSuper.cpp
@@ -35,7 +35,7 @@ int CConnectionSuper::SendChatData(std::vector<std::wstring> vecChatData)
 
 	for (size_t i = 0; i < nVecSize; i++)
 	{
-		int nMsgLength = vecChatData[i].size() * sizeof(wchar_t);
+		int nMsgLength = (int)vecChatData[i].size() * (int)sizeof(wchar_t);
 		::send(m_ConnectionSocket, (const char*)&nMsgLength, sizeof(nMsgLength), 0);
 		::send(m_ConnectionSocket, (const char*)vecChatData[i].c_str(), nMsgLength, 0);
 	}
@@ -44,7 +44,7 @@ int CConnectionSuper::SendChatData(std::vector<std::wstring> vecChatData)
 
 int CConnectionSuper::Send(std::wstring strMessage)
 {
-	int nLength = strMessage.length() * (int)sizeof(wchar_t);
+	int nLength = (int)strMessage.length() * (int)sizeof(wchar_t);
 	::send(m_ConnectionSocket, (const char*)&nLength, sizeof(nLength), 0);
 	::send(m_ConnectionSocket, (const char*)strMessage.c_str(), nLength, 0);
